requirements.txt
pyobjc-core
psycopg2-binary==2.8.6
bleak==0.9.1